backend/requirements.txt
click==8.0.1
Flask
Flask-Cors==3.0.10
itsdangerous==2.0.1
Jinja2
MarkupSafe==2.0.1
pymongo==3.12.0
python-dotenv==0.19.0
PyYAML==5.4.1
six==1.16.0
Werkzeug==2.0.1
gunicorn==20.1.0
flask_monitoringdashboard==3.1.0
